requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
Pillow==9.0.1
email-validator==1.1.3
django-debug-toolbar==3.2.4
djangorestframework-simplejwt==5.1.0
flake8==4.0.1

